fix scrapper to adpat django
