+++ device_problem/models.py
@@ -9,9 +9,9 @@ from django.db import models
 class Maude(models.Model):
     mdr_report_key = models.CharField(max_length=191)
     device_problem = models.CharField(max_length=191)
-    brand_name = models.TextField(blank=True)
-    manufacturer_name = models.TextField(blank=True)
-    device_name = models.TextField(blank=True)
+    brand_name = models.TextField(blank=True, null=True)
+    manufacturer_name = models.TextField(blank=True, null=True)
+    device_name = models.TextField(blank=True, null=True)
 
     class Meta:
         unique_together = ('mdr_report_key', 'device_problem',)
