+++ device_problem/urls.py
@@ -1,0 +1,5 @@
+# from django.conf.urls import url
+# from . import views
+
+urlpatterns = [
+]
