+++ device_problem/views.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
-
+# from django.http import HttpResponse
+# from .device_problem_scrapper import DeviceProblemScrapper
 # from django.shortcuts import render
 
 # Create your views here.
